# Enable fast saving mode for PNGs