Makes saving PNGs 2.5x faster for me at only 2% (!) increase in filesize, although please note that I only did basic testing. Presumably, more compressible images would end up with a bigger filesize penalty.

Helps with Movie Maker, see godotengine/godot-proposals#4751

Fixes #51868