@fire 2023-03-03
I'm interested. Not able to review this week due to a brief break from godot engine 4.0 release.