@Calinou 2023-04-26
## Benchmark

**CPU:** Intel Core i9-13900K
**GPU:** GeForce RTX 4090
**OS:** Fedora 37

Time taken to render https://github.com/Calinou/godot-movie-maker-demo:

Type | AVI | PNG - `master` | PNG - this PR
-|-|-|-
**Editor 1280×720** | 1m07s | 6m25s | 1m41s
**Editor 3840×2160** | 6m49s | 48m00s | 10m44s
**Release template 1280×720** | 39s | 6m36s | 1m48s
**Release template 3840×2160** | 3m31s | 47m06s | 9m37s

File size of first frame in 3840×2160:

PNG - `master` | PNG - this PR
-|-
7,173 KB | 9,432 KB

Considering saving time is the most important criterion in Movie Maker, I'd say this increase in file size is fine. It's common for games in general to take large, unoptimized screenshots to ensure they save quickly (so they don't interrupt the action for too long).