@clayjohn 2023-04-26
Won't this increase the filesize of every PNG saved by the engine? Not just the ones created by the MovieMaker feature?