@fire 2023-04-26
> t's common for games in general to take large, unoptimized screenshots to ensure they save quickly (so they don't interrupt the action for too long).

I think @myaaaaaaaaa is saying that's a good tradeoff.