@myaaaaaaaaa 2023-04-27
> > It's common for games in general to take large, unoptimized screenshots to ensure they save quickly (so they don't interrupt the action for too long).
> 
> I think @myaaaaaaaaa is saying that's a good tradeoff.

@Calinou , not me :slightly_smiling_face: 

> Won't this increase the filesize of every PNG saved by the engine? Not just the ones created by the MovieMaker feature?

The proper solution would definitely be to add a filesize option to `Image::save_png()`. This was more of a quick post to bring the existence of `PNG_IMAGE_FLAG_FAST` to people's attention, but in the form of a PR in case the Godot team wanted to merge it in as an interim solution.