@Calinou 2023-04-27
> Won't this increase the filesize of every PNG saved by the engine? Not just the ones created by the MovieMaker feature?

It will, but the engine defaults to saving lossless WebP when importing images since 3.5.