@PeterMarques 2023-07-01
In the "Marshalls.raw_to_base64(vp.get_texture().get_data().save_png_to_buffer())" that is very slow, and have several bottleneck point, one being the get_texture itself that can be better implemented to have bigger delay but low cpu usage, the Save_png_to_buffer, is the one that consume the most CPU time.