@akien-mga 2023-08-02
> That said, we could make `save_png()` accept a compression mode parameter (uncompressed, fast compression, best compression).

I agree we should make this a flag in `save_png()`. Also discussed it with @reduz who agreed.

Could default to fast compression, but users should still have a way to opt in to best compression if they value size over compression speed.